Refactor codapPhone into multiple files
